New platform agg_types deps (#33372)
* Moved auto_select_if_only_one directive

* Moved validate-date-interval directive

* Moved scroll-bottom directive

* Moved input number directive + Fix validate-date-interval directive

* Moved validate ip and validate cidr mask directives

* Moved validate date math and documentation href directives

* Moved validate agg directive

* Moved click focus directive

* removed validate date interval from autoload
