+++ src/legacy/ui/public/agg_types/__tests__/directives/validate_cidr_mask.js
@@ -19,12 +19,13 @@
 
 import expect from 'expect.js';
 import ngMock from 'ng_mock';
-import '../input_number';
+import '../../directives/validate_cidr_mask';
 
-describe('Number input directive', function () {
+
+describe('Validate CIDR mask directive', function () {
   let $compile;
   let $rootScope;
-  const html = '<input type="text" ng-model="value" input-number />';
+  const html = '<input type="text" ng-model="value" validate-cidr-mask />';
 
   beforeEach(ngMock.module('kibana'));
 
@@ -33,35 +34,71 @@ describe('Number input directive', function () {
     $rootScope = _$rootScope_;
   }));
 
-  it('should allow whole numbers', function () {
+  it('should allow empty input', function () {
     const element = $compile(html)($rootScope);
 
-    $rootScope.value = '123';
+    $rootScope.value = '';
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
 
-    $rootScope.value = '1';
+    $rootScope.value = null;
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
 
-    $rootScope.value = '-5';
+    $rootScope.value = undefined;
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
   });
 
-  it('should allow numbers with decimals', function () {
+  it('should allow valid CIDR masks', function () {
     const element = $compile(html)($rootScope);
 
-    $rootScope.value = '123.0';
+    $rootScope.value = '0.0.0.0/1';
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
 
-    $rootScope.value = '1.2';
+    $rootScope.value = '128.0.0.1/31';
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
 
-    $rootScope.value = '-5.5';
+    $rootScope.value = '1.2.3.4/2';
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
+
+    $rootScope.value = '67.129.65.201/27';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-valid')).to.be.ok();
+  });
+
+  it('should disallow invalid CIDR masks', function () {
+    const element = $compile(html)($rootScope);
+
+    $rootScope.value = 'hello, world';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-invalid')).to.be.ok();
+
+    $rootScope.value = '0.0.0.0';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-invalid')).to.be.ok();
+
+    $rootScope.value = '0.0.0.0/0';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-invalid')).to.be.ok();
+
+    $rootScope.value = '0.0.0.0/33';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-invalid')).to.be.ok();
+
+    $rootScope.value = '256.0.0.0/32';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-invalid')).to.be.ok();
+
+    $rootScope.value = '0.0.0.0/32/32';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-invalid')).to.be.ok();
+
+    $rootScope.value = '1.2.3/1';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-invalid')).to.be.ok();
   });
 });
