+++ src/legacy/ui/public/agg_types/__tests__/directives/validate_ip.js
@@ -19,13 +19,13 @@
 
 import expect from 'expect.js';
 import ngMock from 'ng_mock';
-import '../validate_date_math';
+import '../../directives/validate_ip';
 
 
-describe('Validate date math directive', function () {
+describe('Validate IP directive', function () {
   let $compile;
   let $rootScope;
-  const html = '<input type="text" ng-model="value" validate-date-math />';
+  const html = '<input type="text" ng-model="value" validate-ip />';
 
   beforeEach(ngMock.module('kibana'));
 
@@ -34,55 +34,63 @@ describe('Validate date math directive', function () {
     $rootScope = _$rootScope_;
   }));
 
-  it('should allow valid date math', function () {
+  it('should allow empty input', function () {
     const element = $compile(html)($rootScope);
 
-    $rootScope.value = 'now';
+    $rootScope.value = '';
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
 
-    $rootScope.value = '2012-02-28';
+    $rootScope.value = null;
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
 
-    $rootScope.value = 'now-3d';
-    $rootScope.$digest();
-    expect(element.hasClass('ng-valid')).to.be.ok();
-
-    $rootScope.value = 'now-3M/M';
-    $rootScope.$digest();
-    expect(element.hasClass('ng-valid')).to.be.ok();
-
-    $rootScope.value = '2012-05-31||-3M/M';
+    $rootScope.value = undefined;
     $rootScope.$digest();
     expect(element.hasClass('ng-valid')).to.be.ok();
   });
 
-  it('should disallow invalid date math', function () {
+  it('should allow valid IP addresses', function () {
+    const element = $compile(html)($rootScope);
+
+    $rootScope.value = '0.0.0.0';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-valid')).to.be.ok();
+
+    $rootScope.value = '0.0.0.1';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-valid')).to.be.ok();
+
+    $rootScope.value = '126.45.211.34';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-valid')).to.be.ok();
+
+    $rootScope.value = '255.255.255.255';
+    $rootScope.$digest();
+    expect(element.hasClass('ng-valid')).to.be.ok();
+  });
+
+  it('should disallow invalid IP addresses', function () {
     const element = $compile(html)($rootScope);
 
     $rootScope.value = 'hello, world';
     $rootScope.$digest();
     expect(element.hasClass('ng-invalid')).to.be.ok();
 
-    $rootScope.value = 'now+-5w';
+    $rootScope.value = '0.0.0';
     $rootScope.$digest();
     expect(element.hasClass('ng-invalid')).to.be.ok();
 
-    $rootScope.value = '2012-02-31';
+    $rootScope.value = '256.0.0.0';
     $rootScope.$digest();
     expect(element.hasClass('ng-invalid')).to.be.ok();
 
-    $rootScope.value = '5/5/2005+3d';
+    $rootScope.value = '-1.0.0.0';
     $rootScope.$digest();
     expect(element.hasClass('ng-invalid')).to.be.ok();
-  });
 
-  it('should allow empty values', function () {
-    const element = $compile(html)($rootScope);
-
-    $rootScope.value = '';
+    $rootScope.value = Number.MAX_VALUE;
     $rootScope.$digest();
-    expect(element.hasClass('ng-valid')).to.be.ok();
+    expect(element.hasClass('ng-invalid')).to.be.ok();
   });
 });
