+++ src/legacy/ui/public/agg_types/buckets/date_histogram.js
@@ -21,7 +21,7 @@ import _ from 'lodash';
 import chrome from '../../chrome';
 import moment from 'moment-timezone';
 import '../../filters/field_type';
-import '../../validate_date_interval';
+import '../directives/validate_date_interval';
 import { BucketAggType } from './_bucket_agg_type';
 import { TimeBuckets } from '../../time_buckets';
 import { createFilterDateHistogram } from './create_filter/date_histogram';
