+++ src/legacy/ui/public/agg_types/buckets/filters.js
@@ -23,6 +23,7 @@ import angular from 'angular';
 import { BucketAggType } from './_bucket_agg_type';
 import { createFilterFilters } from './create_filter/filters';
 import { decorateQuery, luceneStringToDsl } from '@kbn/es-query';
+import '../directives/click_focus';
 import filtersTemplate from '../controls/filters.html';
 import { i18n } from '@kbn/i18n';
