+++ src/legacy/ui/public/agg_types/buckets/ip_range.js
@@ -18,8 +18,8 @@
  */
 
 import _ from 'lodash';
-import '../../directives/validate_ip';
-import '../../directives/validate_cidr_mask';
+import '../directives/validate_ip';
+import '../directives/validate_cidr_mask';
 import { BucketAggType } from './_bucket_agg_type';
 import { createFilterIpRange } from './create_filter/ip_range';
 import ipRangesTemplate from '../controls/ip_ranges.html';
