+++ src/legacy/ui/public/agg_types/directives/auto_select_if_only_one.js
@@ -17,25 +17,20 @@
  * under the License.
  */
 
-import _ from 'lodash';
-import $ from 'jquery';
-import { uiModules } from '../modules';
+import { uiModules } from '../../modules';
 const module = uiModules.get('kibana');
 
-module.directive('clickFocus', function () {
+module.directive('autoSelectIfOnlyOne', function () {
   return {
-    scope: {
-      clickFocus: '='
-    },
     restrict: 'A',
-    link: function ($scope, $elem) {
-      function handler() {
-        const focusElem = $.find('input[name=' + $scope.clickFocus + ']');
-        if (focusElem[0]) focusElem[0].focus();
-      }
-
-      $elem.bind('click', handler);
-      $scope.$on('$destroy', _.bindKey($elem, 'unbind', 'click', handler));
+    require: 'ngModel',
+    link: function (scope, element, attributes, ngModelCtrl) {
+      scope.$watchCollection(attributes.autoSelectIfOnlyOne, (options) => {
+        if (options && options.length === 1) {
+          ngModelCtrl.$setViewValue(options[0]);
+          ngModelCtrl.$render();
+        }
+      });
     }
   };
 });
