+++ src/legacy/ui/public/agg_types/directives/input_number.js
@@ -17,16 +17,21 @@
  * under the License.
  */
 
-import { getDocLink } from '../documentation_links';
-import { uiModules } from '../modules';
-
+import { uiModules } from '../../modules';
 const module = uiModules.get('kibana');
 
-module.directive('documentationHref', function () {
+module.directive('inputNumber', function () {
   return {
     restrict: 'A',
-    link: function (scope, element, attributes) {
-      element.attr('href', getDocLink(attributes.documentationHref));
+    require: 'ngModel',
+    link: function ($scope, $elem, attrs, ngModel) {
+      ngModel.$parsers.push(checkNumber);
+      ngModel.$formatters.push(checkNumber);
+
+      function checkNumber(value) {
+        ngModel.$setValidity('number', !isNaN(parseFloat(value)));
+        return value;
+      }
     }
   };
 });
