+++ src/legacy/ui/public/agg_types/directives/scroll_bottom.js
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { uiModules } from '../modules';
+import { uiModules } from '../../modules';
 const module = uiModules.get('kibana');
 
 module.directive('kbnScrollBottom', function () {
