+++ src/legacy/ui/public/agg_types/directives/validate_date_interval.js
@@ -17,9 +17,9 @@
  * under the License.
  */
 
-import { parseInterval } from './utils/parse_interval';
-import { uiModules } from './modules';
-import { leastCommonInterval } from './vis/lib/least_common_interval';
+import { parseInterval } from '../../utils/parse_interval';
+import { uiModules } from '../../modules';
+import { leastCommonInterval } from '../../vis/lib/least_common_interval';
 
 uiModules
   .get('kibana')
