+++ src/legacy/ui/public/agg_types/index.js
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import '../directives/validate_agg';
+import './directives/validate_agg';
 import './agg_params';
 import { IndexedArray } from '../indexed_array';
 import { countMetricAgg } from './metrics/count';
